requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pandas>=2.2
numpy>=1.26
python-docx>=1.1
openpyxl>=3.1
matplotlib>=3.9
python-multipart>=0.0.9
